feat: simplify add pi
